+++ Flakkari/Network/Socket.hpp
@@ -23,6 +23,7 @@
 #include <memory>
 #include <optional>
 #include <utility>
+#include <vector>
 
 #ifdef _WIN32
 typedef SOCKET SOCKET;
@@ -160,7 +161,7 @@ public:
      * @brief Bind the socket to the address and port specified in the constructor.
      *
      */
-    void bind();
+    void bind() const;
 
     /**
      * @brief Start listening for incoming connections.
@@ -168,7 +169,7 @@ public:
      *
      * @param backlog  Maximum number of pending connections. Default is SOMAXCONN.
      */
-    void listen(int backlog = SOMAXCONN);
+    void listen(int backlog = SOMAXCONN) const;
 
     /**
      * @brief Accept an incoming connection.
@@ -176,7 +177,7 @@ public:
      *
      * @return std::shared_ptr<Socket>  The socket of the client.
      */
-    void connect();
+    void connect() const;
 
     /**
      * @brief Disconnect from the server.
@@ -190,14 +191,14 @@ public:
      *
      * @return std::shared_ptr<Socket>  The socket of the client.
      */
-    std::shared_ptr<Socket> accept();
+    std::shared_ptr<Socket> accept() const;
 
     /**
      * @brief Set the socket to blocking or non-blocking.
      *
      * @param blocking  True to set the socket to blocking, false to set it to non-blocking.
      */
-    void setBlocking(bool blocking = true);
+    void setBlocking(bool blocking = true) const;
 
     /**
      * @brief Send data to the socket.
@@ -206,7 +207,7 @@ public:
      * @param data  Data to send.
      * @param flags  Flags to pass to the send function.
      */
-    void send(const Buffer &data, int flags = 0);
+    void send(const Buffer &data, int flags = 0) const;
 
     /**
      * @brief Send data to the socket.
@@ -216,7 +217,7 @@ public:
      * @param size  Size of the data to send.
      * @param flags  Flags to pass to the send function.
      */
-    void send(const Buffer &data, size_t size, int flags = 0);
+    void send(const Buffer &data, size_t size, int flags = 0) const;
 
     /**
      * @brief Send data to the socket.
@@ -226,7 +227,7 @@ public:
      * @param data  Data to send.
      * @param flags  Flags to pass to the send function.
      */
-    void sendTo(const std::shared_ptr<Address> &address, const Buffer &data, int flags = 0);
+    void sendTo(const std::shared_ptr<Address> &address, const Buffer &data, int flags = 0) const;
 
     /**
      * @brief Send data to the socket.
@@ -237,7 +238,7 @@ public:
      * @param size  Size of the data to send.
      * @param flags  Flags to pass to the send function.
      */
-    void sendTo(const std::shared_ptr<Address> &address, const byte *data, const size_t &size, int flags = 0);
+    void sendTo(const std::shared_ptr<Address> &address, const byte *data, const size_t &size, int flags = 0) const;
 
     /**
      * @brief Receive data from the socket.
@@ -247,7 +248,7 @@ public:
      * @param flags  Flags to pass to the recv function.
      * @return std::optional<Buffer>  The data received.
      */
-    std::optional<Buffer> receive(size_t size, int flags = 0);
+    std::optional<Buffer> receive(size_t size, int flags = 0) const;
 
     /**
      * @brief Receive data from the socket. The size of the data is determined by the size of the buffer.
@@ -256,7 +257,7 @@ public:
      * @param flags  Flags to pass to the recv function.
      * @return std::optional<Buffer>  The data received.
      */
-    std::optional<Buffer> receive(int flags = 0);
+    std::optional<Buffer> receive(int flags = 0) const;
 
     /**
      * @brief Receive data from the socket.
@@ -267,7 +268,7 @@ public:
      * @return std::optional<std::pair<std::shared_ptr<Address>, Buffer>>  The data received and the address of the
      * sender.
      */
-    std::optional<std::pair<std::shared_ptr<Address>, Buffer>> receiveFrom(size_t size, int flags = 0);
+    std::optional<std::pair<std::shared_ptr<Address>, Buffer>> receiveFrom(size_t size, int flags = 0) const;
 
     /**
      * @brief Receive data from the socket. The size of the data is determined by the size of the buffer.
@@ -277,13 +278,25 @@ public:
      * @return std::optional<std::pair<std::shared_ptr<Address>, Buffer>>  The data received and the address of the
      * sender.
      */
-    std::optional<std::pair<std::shared_ptr<Address>, Buffer>> receiveFrom(int flags = 0);
+    std::optional<std::pair<std::shared_ptr<Address>, Buffer>> receiveFrom(int flags = 0) const;
+
+    /**
+     * @brief Receive multiple UDP messages in one syscall (batch).
+     * This function is only used by UDP sockets.
+     *
+     * @param maxMessages  Maximum number of messages to receive in one call.
+     * @param flags  Flags to pass to recvfrom / recvmmsg.
+     * @return vector of (Address, Buffer) pairs received. Empty if none or on EAGAIN/EWOULDBLOCK.
+     * @note On Linux this uses recvmmsg for efficiency. On other platforms it falls back to repeated recvfrom.
+     */
+    std::vector<std::pair<std::shared_ptr<Address>, Buffer>> receiveBatch(uint32_t maxMessages = 16u,
+                                                                          int flags = 0) const;
 
     /**
      * @brief Close the socket.
      *
      */
-    void close();
+    void close() const;
 
     /**
      * @brief Convert Socket to string (ip:port).
@@ -292,9 +305,17 @@ public:
      */
     operator std::string() const;
 
+private:
+    /**
+     * @brief Get the Ip Type From Family object
+     *
+     * @param addrStorage  The sockaddr_storage to get the family from
+     * @return Address::IpType  The IpType
+     */
+    [[nodiscard]] inline Address::IpType getIpTypeFromFamily(const sockaddr_storage &addrStorage) const;
+
 protected:
 private:
-    std::mutex _mutex;
     socket_t _socket;
     std::shared_ptr<Address> _address = nullptr;
 };